# chapter5/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Root",
                "operationId": "root__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/v0/players/": {
            "get": {
                "summary": "Read Players",
                "operationId": "read_players_v0_players__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "minimum_last_changed_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "title": "Minimum Last Changed Date"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "First Name"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Last Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Player"
                                    },
                                    "title": "Response Read Players V0 Players  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/players/{player_id}": {
            "get": {
                "summary": "Read Player",
                "operationId": "read_player_v0_players__player_id__get",
                "parameters": [
                    {
                        "name": "player_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Player Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/performances/": {
            "get": {
                "summary": "Read Performances",
                "operationId": "read_performances_v0_performances__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "minimum_last_changed_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "title": "Minimum Last Changed Date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Performance"
                                    },
                                    "title": "Response Read Performances V0 Performances  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/leagues/{league_id}": {
            "get": {
                "summary": "Read League",
                "operationId": "read_league_v0_leagues__league_id__get",
                "parameters": [
                    {
                        "name": "league_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "League Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/League"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/leagues/": {
            "get": {
                "summary": "Read Leagues",
                "operationId": "read_leagues_v0_leagues__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "minimum_last_changed_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "title": "Minimum Last Changed Date"
                        }
                    },
                    {
                        "name": "league_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "League Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/League"
                                    },
                                    "title": "Response Read Leagues V0 Leagues  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/teams/": {
            "get": {
                "summary": "Read Teams",
                "operationId": "read_teams_v0_teams__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "minimum_last_changed_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "title": "Minimum Last Changed Date"
                        }
                    },
                    {
                        "name": "team_name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Team Name"
                        }
                    },
                    {
                        "name": "league_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "title": "League Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Team"
                                    },
                                    "title": "Response Read Teams V0 Teams  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v0/counts/": {
            "get": {
                "summary": "Get Count",
                "operationId": "get_count_v0_counts__get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Counts"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Counts": {
                "properties": {
                    "league_count": {
                        "type": "integer",
                        "title": "League Count"
                    },
                    "team_count": {
                        "type": "integer",
                        "title": "Team Count"
                    },
                    "player_count": {
                        "type": "integer",
                        "title": "Player Count"
                    }
                },
                "type": "object",
                "required": [
                    "league_count",
                    "team_count",
                    "player_count"
                ],
                "title": "Counts",
                "description": "Pure response object \u2014 no database table behind it, so no from_attributes."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "League": {
                "properties": {
                    "league_id": {
                        "type": "integer",
                        "title": "League Id"
                    },
                    "league_name": {
                        "type": "string",
                        "title": "League Name"
                    },
                    "scoring_type": {
                        "type": "string",
                        "title": "Scoring Type"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    },
                    "teams": {
                        "items": {
                            "$ref": "#/components/schemas/TeamBase"
                        },
                        "type": "array",
                        "title": "Teams",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "league_id",
                    "league_name",
                    "scoring_type",
                    "last_changed_date"
                ],
                "title": "League",
                "description": "League WITH teams \u2014 used for /v0/leagues/ endpoint."
            },
            "Performance": {
                "properties": {
                    "performance_id": {
                        "type": "integer",
                        "title": "Performance Id"
                    },
                    "player_id": {
                        "type": "integer",
                        "title": "Player Id"
                    },
                    "week_number": {
                        "type": "string",
                        "title": "Week Number"
                    },
                    "fantasy_points": {
                        "type": "number",
                        "title": "Fantasy Points"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    }
                },
                "type": "object",
                "required": [
                    "performance_id",
                    "player_id",
                    "week_number",
                    "fantasy_points",
                    "last_changed_date"
                ],
                "title": "Performance"
            },
            "Player": {
                "properties": {
                    "player_id": {
                        "type": "integer",
                        "title": "Player Id"
                    },
                    "gsis_id": {
                        "type": "string",
                        "title": "Gsis Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "position": {
                        "type": "string",
                        "title": "Position"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    },
                    "performances": {
                        "items": {
                            "$ref": "#/components/schemas/Performance"
                        },
                        "type": "array",
                        "title": "Performances",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "player_id",
                    "gsis_id",
                    "first_name",
                    "last_name",
                    "position",
                    "last_changed_date"
                ],
                "title": "Player",
                "description": "Player WITH performances \u2014 used for /v0/players/ endpoints."
            },
            "PlayerBase": {
                "properties": {
                    "player_id": {
                        "type": "integer",
                        "title": "Player Id"
                    },
                    "gsis_id": {
                        "type": "string",
                        "title": "Gsis Id"
                    },
                    "first_name": {
                        "type": "string",
                        "title": "First Name"
                    },
                    "last_name": {
                        "type": "string",
                        "title": "Last Name"
                    },
                    "position": {
                        "type": "string",
                        "title": "Position"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    }
                },
                "type": "object",
                "required": [
                    "player_id",
                    "gsis_id",
                    "first_name",
                    "last_name",
                    "position",
                    "last_changed_date"
                ],
                "title": "PlayerBase",
                "description": "Player WITHOUT performances \u2014 used when embedding in Team responses."
            },
            "Team": {
                "properties": {
                    "league_id": {
                        "type": "integer",
                        "title": "League Id"
                    },
                    "team_id": {
                        "type": "integer",
                        "title": "Team Id"
                    },
                    "team_name": {
                        "type": "string",
                        "title": "Team Name"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    },
                    "players": {
                        "items": {
                            "$ref": "#/components/schemas/PlayerBase"
                        },
                        "type": "array",
                        "title": "Players",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "league_id",
                    "team_id",
                    "team_name",
                    "last_changed_date"
                ],
                "title": "Team",
                "description": "Team WITH players \u2014 used for /v0/teams/ endpoint."
            },
            "TeamBase": {
                "properties": {
                    "league_id": {
                        "type": "integer",
                        "title": "League Id"
                    },
                    "team_id": {
                        "type": "integer",
                        "title": "Team Id"
                    },
                    "team_name": {
                        "type": "string",
                        "title": "Team Name"
                    },
                    "last_changed_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Last Changed Date"
                    }
                },
                "type": "object",
                "required": [
                    "league_id",
                    "team_id",
                    "team_name",
                    "last_changed_date"
                ],
                "title": "TeamBase",
                "description": "Team WITHOUT players \u2014 used when embedding in League responses."
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}